Small Backend Bug and Frontend Fixes
styling implemented.  Letterbox no longer showcases an issue.  Finally (FINALLY) Got the keyboard to update properly.  It was an issue with immutability.  During the dictionary update, I was using forEach.  Also added useRef to allow any key handling typed to be added into the board without the user clicking on it. Up next is style and animation.